# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RomaniaDate Dating App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/discover:
    get:
      operationId: discoverUsers
      tags: [users]
      summary: Get users to discover (for swiping)
      parameters:
        - name: city
          in: query
          schema:
            type: ["string", "null"]
        - name: minAge
          in: query
          schema:
            type: ["integer", "null"]
        - name: maxAge
          in: query
          schema:
            type: ["integer", "null"]
        - name: onlineOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/report:
    post:
      operationId: reportUser
      tags: [users]
      summary: Report a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "200":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{userId}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/who-liked-me:
    get:
      operationId: getWhoLikedMe
      tags: [users]
      summary: Get users who liked me (VIP only)
      responses:
        "200":
          description: Users who liked me
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
        "403":
          description: VIP required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/settings:
    put:
      operationId: updateSettings
      tags: [users]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/vip:
    post:
      operationId: activateVip
      tags: [users]
      summary: Activate VIP subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateVipRequest"
      responses:
        "200":
          description: VIP activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        birthDate:
          type: string
        age:
          type: integer
        ageCategory:
          type: string
          enum: [teen, adult]
        gender:
          type: string
          enum: [boy, girl]
        preference:
          type: string
          enum: [boy, girl, both]
        city:
          type: string
        bio:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        diamonds:
          type: integer
        isVip:
          type: boolean
        vipExpiresAt:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isPrivate:
          type: boolean
        isBoosted:
          type: boolean
        boostExpiresAt:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
        rank:
          type: ["string", "null"]
        referralCode:
          type: string
        isOnline:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        likesGivenToday:
          type: integer
        superlikesGivenToday:
          type: integer
        notifyNewMatch:
          type: boolean
        notifyNewMessage:
          type: boolean
        notifyDiamonds:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - birthDate
        - age
        - ageCategory
        - gender
        - preference
        - city
        - photos
        - interests
        - diamonds
        - isVip
        - isVerified
        - isPrivate
        - isBoosted
        - level
        - xp
        - referralCode
        - isOnline
        - isAdmin
        - likesGivenToday
        - superlikesGivenToday
        - notifyNewMatch
        - notifyNewMessage
        - notifyDiamonds
        - createdAt
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        ageCategory:
          type: string
          enum: [teen, adult]
        gender:
          type: string
          enum: [boy, girl]
        city:
          type: string
        bio:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        isBoosted:
          type: boolean
        isOnline:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        level:
          type: integer
        rank:
          type: ["string", "null"]
      required:
        - id
        - name
        - age
        - ageCategory
        - gender
        - city
        - photos
        - interests
        - isVerified
        - isBoosted
        - isOnline
        - level
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        photos:
          type: ["array", "null"]
          items:
            type: string
        interests:
          type: ["array", "null"]
          items:
            type: string
        city:
          type: ["string", "null"]
    UpdateSettingsRequest:
      type: object
      properties:
        city:
          type: ["string", "null"]
        preference:
          type: ["string", "null"]
          enum: [boy, girl, both, null]
        isPrivate:
          type: ["boolean", "null"]
        notifyNewMatch:
          type: ["boolean", "null"]
        notifyNewMessage:
          type: ["boolean", "null"]
        notifyDiamonds:
          type: ["boolean", "null"]
...
    ReportRequest:
      type: object
      properties:
        reason:
          type: string
          enum: [spam, inappropriate, fake, harassment, other]
        description:
          type: ["string", "null"]
      required:
        - reason
...
    ActivateVipRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [monthly, quarterly, yearly]
      required:
        - plan