# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RomaniaDate Dating App API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUserAgeCategory } from "./publicUserAgeCategory";
import type { PublicUserGender } from "./publicUserGender";

export interface PublicUser {
  id: number;
  name: string;
  age: number;
  ageCategory: PublicUserAgeCategory;
  gender: PublicUserGender;
  city: string;
  /** @nullable */
  bio?: string | null;
  photos: string[];
  interests: string[];
  isVerified: boolean;
  isBoosted: boolean;
  isOnline: boolean;
  /** @nullable */
  lastSeen?: string | null;
  level: number;
  /** @nullable */
  rank?: string | null;
}